Started working on UI and getting basic UI Structure down

+++ app/src/main/java/com/msruback/maxixe/database/entities/Contact.kt
@@ -4,7 +4,7 @@ import androidx.room.*
 
 @Entity("Contacts")
 data class Contact(
-    @PrimaryKey(true) @ColumnInfo("Id") val id: Int,
+    @PrimaryKey(true) @ColumnInfo("Id") val id: Long,
     @ColumnInfo("Name") var name: String,
     @ColumnInfo("Pronouns") var pronouns: String,
     @ColumnInfo("Desc") var desc: String,
